PiperOrigin-RevId: 549316369

+++ it/google-cloud-platform/src/main/java/com/google/cloud/teleport/it/gcp/spanner/SpannerResourceManager.java
@@ -351,7 +351,11 @@ public final class SpannerResourceManager implements ResourceManager {
   public synchronized void cleanupAll() {
     try {
       LOG.info("Deleting instance {}...", instanceId);
-      instanceAdminClient.deleteInstance(instanceId);
+
+      if (instanceAdminClient != null) {
+        instanceAdminClient.deleteInstance(instanceId);
+      }
+
       hasInstance = false;
       hasDatabase = false;
     } catch (SpannerException e) {
